move navbar to component
