+++ pages/blog/[id].js
@@ -1,4 +1,4 @@
-import Layout from '../../components/layout.js'
+import Layout from '../../src/components/layout.js'
 import { getPostIDs, getPostData } from '../../lib/posts.js'
 import styles from './blog.module.css';
